Debug Overflow interrupt (avr) WARNING : will have to change all interrupts vector. Debug Stream binary mode.

+++ include/sasiae/device/stream/eeprom_stream.hpp
@@ -19,7 +19,7 @@ char EepromStream::getValue(void) {
 
 void EepromStream::setValue(char val) {
   std::ostringstream oss;
-  oss << "value " << (int)val;
+  oss << "value " << (int)((u8)val);
 
   ClientThread::instance().
     sendDeviceMessage(*this,
